[ffwizard] fix bug collectd does not start
Issue freifunk-berlin/firmware#163: We can not disable collectd and rrdtool
because there is no option in the ui to enable them again. Also the
luci_service should not be disabled even if the user does not enable
statistics during wizard setup because the luci_statistics ui needs the
service to be enabled to work.