[ffwizard] set start and limit in dhcp configuration
We should not use the openwrt defaults for start and limit (100, 150). This
leads to wrong values for the dhcp range and a suboptimal usage of the ip space.

We start at an offset of 2. The first address after the network address is used
by the router. We limit the number of addresses to the number of host addresses
in the subnet minus the one address for the router.

This should fix:

https://github.com/freifunk-berlin/firmware/issues/311