dhcp-defaults: don't announce as default-gw
By default we will not announce our self as default-gw to the
clients, to prevent them sending us pakets to forward, but not
having a configured uplink yet.
So the client can be connected to his fully working network
(via WiFI) and connect to us for configuration (via LAN),
without breaking it's working uplink connection.
As we do this on "LAN"-interface, which will be deleted by the
wizard this will not cause problems on regular installs.
Only systems configured manually and keeping the LAN-interface,
have to take care (esp. backbone-flavor).

this implements https://github.com/freifunk-berlin/firmware-packages/pull/123