luci-app-owm: remove obviously dead code
With the invention of git we don't want maintain possibly usefull code
in comments inside the source file. If we need the dead code later on we
can check the git history.

The removed code includes debug print messages and code that looks up
the IPv4+6 default gateway and adds it to the owm data. For some reason
the code does not work.