migration: run as 1st of freifunk-berlin scripts
As of the renaming of the guard-setting for VPN03
(vpn03_rename_guard) the migration hat to be run before
any of the relevant configs are applied to the system.
On fresh installed systems the migration will no do anything,
so it should have no effect here. On systems being upgraded
the config-changes are protected by the "guard"-function and
will finish w/o doing any action, so it seems also to have no
sideeffect here.