firewall-defaults, wizard: add separate zone ffvpn
this zone should be used for traffic to any uplink- / VPN-connection.
This traffic if not really freifunk-traffic and should be considered different.

Put the default interface for the VPN/uplink into the correct firewall-zone by default.
As the ffvpn-interface is in the correct ffvpn-zone by default, this setting don't need to be done by the wizard anymore.