uplink-vpn03,tunnelberlin: explicitly define IP-Masquerading setting
As VPN03 is requesting to don't do IP-masquerading on the router, but
other uplinks (e.g. community-tunnel Berlin) relying on this, we setup
this feature within every uplink-profile. This ensures it's correct
for the corresponding tunnel.
To allow easy switching between different tunnels, this setting is
not protected by the guard-feature.