luci-app-owm: change field names for arptable in owm.lua
With a commit in the luci code base the function luci.sys.net.arptable
was replaced by luci.ip.neighbors:

https://github.com/openwrt/luci/commit/366707a681459a4d520dc97024ea0a4b3c24a326

The field names of the returned table are different to the old field
names. Therefore we have to change the field names. The values of the
table are of type userdata and we have to convert to string with the
tostring method.