uplink macddr: store the macaddr of ffuplink_dev in ffberlin-uplink
The macaddr is saved in ffberlin-uplink.uplink.macaddr.  In this was it
will also be reused when changing uplink types back to a type which
previously had a macaddr.  This is important for those who set up special
mac filtering on their uplink routers (for example guest network or static
DHCP leases).

Situation:
install no-tunnel
change to openvpn
change back to no-tunnel

This addresses issue https://github.com/freifunk-berlin/firmware/issues/639