mirgation: add OLSRd core dump garbage collectiong
the OLSR daemon in the 1.0.x firmware dumps a core whenever it
crashes.  This can slowly kill a router as it runs out of memory.  For the
1.0.x firmwares, the function r1_0_2_add_olsrd_garbage_collection has been
added

the OLSR daemon for the 1.1.x firmware does not have this issue.  For 1.1.x
firwares, the function r1_1_0_remove_olsrd_garbage_collection to undo the
garbage collection.

The garbage collection is done as a cron job which runs daily at 04:23
23 4 * * *	rm -f /tmp/olsrd*core

This addresses issue https://github.com/freifunk-berlin/firmware/issues/642