wizard: fix hasRootPass by implementing check via luci.sys.user.getpasswd
The first boot of the routers had stopped going to the changePassword
dialog.  After further checking, the function hasRootPass was timing out
at the call to wget, and resulting in a false positive.

The luci method to check if a password has been set is to use
luci.sys.user.getpasswd("root").  This method is fast and relies on
system calls instead of external applications such as wget.