network-defaults: add pingcheck support
The scripts, when installed with pingcheck, will set routing rules
for locally generated traffic depending on if the wan interface has
connectivity to the internet.  When the wan interface looses
connectivity, new ip rules are put in place to put the olsr-tunnel
and olsr-default tables with a higher priority than the main table.
When connectivity returns, the rules are removed.

The pingcheck package is now a dependency of freifunk-berlin-network-defaults