add freifunk-berlin-generic.mk to reuse LUCI build-support
The luci.mk file of the LUCI-feed provides some usefull macros
and definitions to build (luci based) packages. These macros
include SRCDIET, JSMIN, CSSTIDY and automatic build of translations.
This script checks the root of every feed for the luci.mk file and
includes it. It errors to logfile, in case the luci.mk is not found
exactly one time.
This way we can use this file from the Luci-feed without linking to
it directly in a possible unknown destination.