wizard: use static file as ffwizard uci-config
Drop creation of ffwizard uci-config from lua-code and use a template-file
included into the package.
Installing this files on a blank node, just unpacks the file from the package,
as it would have been done by the lua code before. On a upgrade of the node,
this file gets unpacked to, but afterwards overwritten by the version from the
sysupgrade-archive afterwards.