ffwizard: enable the ffuplink interface when sharenet is selected
The ffuplink interface is disabled by default on a fresh install.
During the course of running the wizard, the user must select if
they want to share their internet connection or not.  If the user
decides to share their internet, then the ffuplink interface shall
be enabled (setting disabled=0).

This addresses https://github.com/freifunk-berlin/firmware/issues/603