migration: update to the most recent profiles and freifunk defaults
Upon upgrade, the files /etc/config/profile_* and /etc/config/freifunk
are not automatically upgraded to the newest version.  This commit
pulls the profile_* out of the rom partition and overwrites the
profiles in /etc/config.

Additionally the freifunk config is handeled seperately.  This config
file is a mixure of defaults with the addition of the 'contact' and
'community' sections which are configured for the running system.
In this case, the version from rom is copied into /etc/config and
the 'contact' and 'community' sections from the old setup
are imported.

This fixes https://github.com/freifunk-berlin/firmware/issues/641
This also adresses https://github.com/freifunk-berlin/firmware-packages/pull/189#issuecomment-502814486
and https://github.com/freifunk-berlin/firmware-packages/pull/189#issuecomment-502855417