configtools/sharenet-switch: use lua to detect physical interfaces
In Lede-17.01 using the uci-shell interface returned the l3-interface in
"uciget network.*.ifname", which might be the same as the l2-interface. The
l2-devices was avail in "uciget network.*.devices".
On a NSM2 this kind of output was given (vanilla lede-17.01):
  checking interface loopback
   ifnames:lo
   devices:lo
  checking interface lan
   ifnames:br-lan
   devices:eth0
  checking interface wan
   ifnames:eth1
   devices:eth1
  checking interface wan6
   ifnames:eth1
   devices:eth1

Since OpenWrt-18.06 the devices-section is always empty / not present and
ifnames stays the same. So there is no way to get the l2-device via shell.
On a NSM2 this kind of output was given (vanilla openwrt-19.07):
  checking interface loopback
   ifnames:lo
   devices:
  checking interface lan
   ifnames:br-lan
   devices:
  checking interface wan
   ifnames:eth0
   devices:
  checking interface wan6
   ifnames:eth0
   devices:

Using libuci-lua returns the l2-device in the ifnames-section from
lede-17.01 up to OpenWrt-master (pre OpenWrt 20.xx release currently)
On a NSM2 this kind of output was given (vanilla openwrt-19.07):
  checking interface loopback
   ifnames:lo
   devices:nil
  checking interface lan
   ifnames:eth1
   devices:nil
  checking interface wan
   ifnames:eth0
   devices:nil
  checking interface wan6
   ifnames:eth0
   devices:nil

If this is an intended API-change was asked on the OpenWrt-devel list by:
http://lists.infradead.org/pipermail/openwrt-devel/2020-April/022808.html

the following code was used to compare the results:
--------------------------------------------------

. /lib/functions.sh

swap_port_physical() {
  local port1=$1
  local port2=$2

  is_interface_of() {
    local result_ifname
    local result_device
    config_get result_ifname $1 ifname
    config_get result_device $1 device # found on bridges (ifname is br-dhcp)
    echo "checking interface $1"
    echo " ifnames:$result_ifname"
    echo " devices:$result_device"
  }

  is_interface_of_lua() {

lua <<EOF
uci=require("uci")
x=uci.cursor()
x:foreach("network", "interface", function(s)
 print("checking interface " .. tostring(s[".name"]))
 print(" ifnames:" .. tostring(s["ifname"]))
 print(" devices:" .. tostring(s.device))
end)
EOF

}

  echo "----------- reading by shell -----------"
  local interface
  config_load "network"
  config_foreach is_interface_of "interface"
  echo "----------- reading by lua -----------"
  is_interface_of_lua
}

swap_port_physical eth0
-------------------------------------