system-defaults: set system.was_development flag when running non-release code
Set the system.was_development flag when the system is running non-released code /
development code. This setting will be kept even when a released firmware will be
installed.
This can be used to detect / debug problems that might come up when the system
was running a development-version which did "problematic" changes.