drop "freifunk-berlin-wizard2"
Accoring to issue #746 (https://github.com/freifunk-berlin/firmware/issues/746) it was agreed
that there is nobody working on this wizard anymore and the code is unmaintained for some
time already. So we agreed on removing the code and closing all isses relating to it.

If someone is willing to work on this code again, we can easyily revert this removal.