luci-app-ffwizard-berlin: fix insufficient permissions
The app tries to access the freifunk community files that are placed
at /etc/config/profile_X. If the lua script tries to collect the names
it will get an "permission denied" error. The PR gives the lua script
the permissions to access the config files.

Further, the app access and writes other config files. In addition,
the script executes ffwizard-berlini.

Signed-off-by: Nick Hainke <vincent@systemli.org>