disable ffuplink-interface during initial setup (#185)
* network-defaults: disable ffuplink by default

On a fresh installed system, there is no valid setup of ffuplink. So
we disable this interface, till it's configured by some uplink-preset.

* uplinks: enable ffuplink interface after setup

The default of this interface is disabled, so we must enable it
after it's setup.

* ffwizard: enable the ffuplink interface when sharenet is selected

The ffuplink interface is disabled by default on a fresh install.
During the course of running the wizard, the user must select if
they want to share their internet connection or not.  If the user
decides to share their internet, then the ffuplink interface shall
be enabled (setting disabled=0).

This addresses https://github.com/freifunk-berlin/firmware/issues/603

* Revert "ffwizard: enable the ffuplink interface when sharenet is selected"

This reverts commit 1aebe4530d8d583890901833bcf25ecdf3426008, as it puts code
related to "shareInternet-Settings" in the wireless.lua-file.

* ffwizard: enable ffuplink-interface when "share-internet" was selected

When the user is on the "shareInternet"-page he must have selected to share
his uplink. As he wants to share the uplink, we enable this interface here.

* Revert "uplinks: enable ffuplink interface after setup"

This reverts commit a54856802ebcda02cb37e4eac6d8295b055b4d0c, as we enable the
ffuplink-interface in the ffwizard when reachjing the "sharenet"-page. On
manually configured routers this interface must be enabled in the configuration.

* network-defaults: revert version-bump

this is not needed anymore, sine we use git-autoversioning on master.

Co-authored-by: pmelange <isprotejesvalkata@gmail.com>