wizard: use "redirect_landingurl" feature of mod-freifunk to bring up wizard on 1st-boot
The luci-mod-freifunk has the new feature to redirect to a custom page when LuCI is called
without an explicit page (https://github.com/freifunk/openwrt-packages/commit/21e69cd27d736f13295e2af79108617ccc694b45). Let's
use this to redirect to the wizard and remove the redirection again when the wizard has
finished.
This way we don't need the "redirect on firstboot" patch (up to Kathleen 0.1.2) and can
also drop our forked "luci-mod-freifunk-ui", which is basically the same code with the
mentioned patch included.

Signed-off-by: Sven Roederer <devel-sven@geroedel.de>